pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]
authors = [{name = "Alexandr A.", email = "mail@example.com"}]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]
